+++ test_utils/factories.py
@@ -25,6 +25,7 @@ from enterprise.models import (
     EnterpriseCustomerIdentityProvider,
     EnterpriseCustomerInviteKey,
     EnterpriseCustomerReportingConfiguration,
+    EnterpriseCustomerSsoConfiguration,
     EnterpriseCustomerUser,
     LearnerCreditEnterpriseCourseEnrollment,
     LicensedEnterpriseCourseEnrollment,
@@ -1035,3 +1036,26 @@ class EnterpriseCustomerApiCredentialsFactory(factory.django.DjangoModelFactory)
     name = factory.LazyAttribute(lambda x: FAKER.company())
     client_id = factory.LazyAttribute(lambda x: FAKER.word())
     client_secret = factory.LazyAttribute(lambda x: FAKER.word())
+
+
+class EnterpriseCustomerSsoConfigurationFactory(factory.django.DjangoModelFactory):
+    """
+    EnterpriseCustomerSsoConfiguration factory.
+
+    Creates an instance of EnterpriseCustomerSsoConfiguration with minimal boilerplate.
+    """
+
+    class Meta:
+        """
+        Meta for EnterpriseCustomerSsoConfiguration.
+        """
+
+        model = EnterpriseCustomerSsoConfiguration
+
+    uuid = factory.LazyAttribute(lambda x: UUID(FAKER.uuid4()))
+    enterprise_customer = factory.SubFactory(EnterpriseCustomerFactory)
+    active = True
+    identity_provider = factory.LazyAttribute(lambda x: FAKER.name())
+    metadata_url = factory.LazyAttribute(lambda x: FAKER.url())
+    entity_id = factory.LazyAttribute(lambda x: FAKER.url())
+    update_from_metadata = True